refactor(transformer): encapsulate TypeScript semantic cleanup
