+++ crates/oxc_transformer/src/lib.rs
@@ -60,7 +60,7 @@ use jsx::Jsx;
 use regexp::RegExp;
 use rustc_hash::FxHashMap;
 use state::TransformState;
-use typescript::{RemovedTypeScriptSemantics, TypeScript};
+use typescript::TypeScript;
 
 use crate::plugins::Plugins;
 pub use crate::{
@@ -211,10 +211,10 @@ impl<'a> Transformer<'a> {
 
         let mut reusable_ctx = ReusableTraverseCtx::new(self.state, scoping, allocator);
         traverse_mut_with_ctx(&mut transformer, program, &mut reusable_ctx);
-        let removed_semantics =
-            transformer.x0_typescript.as_mut().map(TypeScript::take_removed_semantics);
         let (mut state, mut scoping) = reusable_ctx.into_state_and_scoping();
-        update_removed_ambient_references(removed_semantics, &mut scoping);
+        if let Some(typescript) = &mut transformer.x0_typescript {
+            typescript.finalize_semantics(&mut scoping);
+        }
         let helpers_used = state.helper_loader.used_helpers.drain().collect();
         let mut diagnostics = react_compiler_diagnostics;
         diagnostics.extend(state.take_errors());
@@ -242,53 +242,6 @@ impl<'a> Transformer<'a> {
         let scoping =
             SemanticBuilder::new().with_enum_eval(true).build(program).semantic.into_scoping();
         (scoping, result.diagnostics)
-    }
-}
-
-fn update_removed_ambient_references(
-    removed_semantics: Option<RemovedTypeScriptSemantics<'_>>,
-    scoping: &mut Scoping,
-) {
-    let Some(RemovedTypeScriptSemantics { mut declarations }) = removed_semantics else {
-        return;
-    };
-    declarations.sort_unstable_by_key(|declaration| {
-        (declaration.symbol_id.index(), declaration.span.start, declaration.span.end)
-    });
-
-    let mut spans = Vec::new();
-    let mut removed_bindings = Vec::new();
-    let mut start = 0;
-    while start < declarations.len() {
-        let symbol_id = declarations[start].symbol_id;
-        let name = declarations[start].name;
-        let mut end = start + 1;
-        while end < declarations.len() && declarations[end].symbol_id == symbol_id {
-            end += 1;
-        }
-
-        spans.clear();
-        spans.extend(declarations[start..end].iter().map(|declaration| declaration.span));
-        if scoping.remove_symbol_declarations(symbol_id, &spans) {
-            start = end;
-            continue;
-        }
-
-        let scope_id = scoping.symbol_scope_id(symbol_id);
-        let outer_symbol_id = scoping
-            .scope_parent_id(scope_id)
-            .and_then(|parent_scope_id| scoping.find_binding(parent_scope_id, name));
-        removed_bindings.push((symbol_id, name, scope_id, outer_symbol_id));
-        start = end;
-    }
-
-    for (symbol_id, name, scope_id, outer_symbol_id) in removed_bindings {
-        scoping.remove_binding(scope_id, name);
-        if let Some(outer_symbol_id) = outer_symbol_id {
-            scoping.rebind_symbol_references(symbol_id, outer_symbol_id);
-        } else {
-            scoping.unresolve_symbol_references(name, symbol_id);
-        }
     }
 }
